Added a timer to skip errors when waking up.
Sometimes when the computer wakes up (from sleep/hibernation, or screen saver) the initial call to twitch may fail.  Since the failure is most likely due to the internet connection rather than a problem with our application or twitch, this is a simple fix/hack to just hide those errors.

We simply hide any errors from the time we wake up for (CHECK_USERS_TIMER_INTERVAL / 2).  Which is 30 seconds as of now.

I also corrected the spelling of ConvertJsonStringToW function.