Show what game people are playing.
When someone comes online, we get a system tray notification bubble showing us the name of the user who came online, along with the game they are playing.  However, after they are online, there's no indication which game people are playing.

Added to the Users menu will be if the user is online, it will show an indented title of what they are playing.

handmade_hero
    Playing: Creative

TODO(bk):  Play around with the size of the buffer.  If the game length is the max length, the menu will expand for the full length.  It looks weird.  Is this what is wanted?