Change UnhandledException to handle Unicode.
Apparently I was not paying attention, or surprised that gcc did not throw a warning/error when trying to assign a wide string literal to a char.  The original intention of the implementation of the function was that it did not need to support unicode characters.  I've changed it so that it can.

Aside:  I was able to generate a dump file when I received an unhandled exception, but now need to work out how to get it working with gcc.  Kind of an oversight...